+++ src/js/discussion/camps/AddCamp.js
