+++ src/js/discussion/camps/Camp.js
