+++ src/js/discussion/camps/CampList.js
