+++ src/js/discussion/camps/CampParent.js
