+++ src/js/discussion/camps/getCampsReducer.js
