+++ src/js/discussion/comments/Counter.js
