+++ src/js/discussion/discuss/Discussion.js
